+++ packages/manager/modules/emailpro/src/dashboard/emailpro-tabs-controller.js
