+++ packages/manager/modules/emailpro/src/dashboard/emailpro-toolbox-groups.controller.js
