+++ packages/manager/modules/emailpro/src/dashboard/emailpro-toolbox-resources.controller.js
