+++ packages/manager/modules/emailpro/src/dashboard/emailpro.constants.js
