+++ packages/manager/modules/emailpro/src/dashboard/emailpro.controller.js
