+++ packages/manager/modules/emailpro/src/dashboard/emailpro.service.js
