+++ packages/manager/modules/emailpro/src/dashboard/emailpro.template.js
@@ -1,51 +1,51 @@
-import emailproRemoveTemplate from './remove/emailpro-remove.html';
-import emailproInformationTemplate from './information/emailpro-information.html';
-import emailproDisclaimerRemoveTemplate from './disclaimer/remove/emailpro-disclaimer-remove.html';
-import emailproDisclaimerUpdateTemplate from './disclaimer/update/emailpro-disclaimer-update.html';
-import emailproDisclaimerTemplate from './disclaimer/emailpro-disclaimer.html';
-import emailproDisclaimerAddTemplate from './disclaimer/add/emailpro-disclaimer-add.html';
-import emailproUpdateRenewTemplate from './update/renew/emailpro-update-renew.html';
-import emailDomainEmailRedirectionUpdateTemplate from './redirection/update/email-domain-email-redirection-update.html';
-import emailDomainEmailRedirectionDeleteTemplate from './redirection/delete/email-domain-email-redirection-delete.html';
-import emailproRedirectionTemplate from './redirection/emailpro-redirection.html';
-import emailDomainEmailRedirectionCreateTemplate from './redirection/create/email-domain-email-redirection-create.html';
-import emailproExternalContactRemoveTemplate from './external-contact/remove/emailpro-external-contact-remove.html';
-import emailproExternalContactUpdateTemplate from './external-contact/update/emailpro-external-contact-update.html';
-import emailproExternalContactAddTemplate from './external-contact/add/emailpro-external-contact-add.html';
-import emailproExternalContactTemplate from './external-contact/emailpro-external-contact.html';
+import emailproRemoveTemplate from '../remove/emailpro-remove.html';
+import emailproInformationTemplate from '../information/emailpro-information.html';
+import emailproDisclaimerRemoveTemplate from '../disclaimer/remove/emailpro-disclaimer-remove.html';
+import emailproDisclaimerUpdateTemplate from '../disclaimer/update/emailpro-disclaimer-update.html';
+import emailproDisclaimerTemplate from '../disclaimer/emailpro-disclaimer.html';
+import emailproDisclaimerAddTemplate from '../disclaimer/add/emailpro-disclaimer-add.html';
+import emailproUpdateRenewTemplate from '../update/renew/emailpro-update-renew.html';
+import emailDomainEmailRedirectionUpdateTemplate from '../redirection/update/email-domain-email-redirection-update.html';
+import emailDomainEmailRedirectionDeleteTemplate from '../redirection/delete/email-domain-email-redirection-delete.html';
+import emailproRedirectionTemplate from '../redirection/emailpro-redirection.html';
+import emailDomainEmailRedirectionCreateTemplate from '../redirection/create/email-domain-email-redirection-create.html';
+import emailproExternalContactRemoveTemplate from '../external-contact/remove/emailpro-external-contact-remove.html';
+import emailproExternalContactUpdateTemplate from '../external-contact/update/emailpro-external-contact-update.html';
+import emailproExternalContactAddTemplate from '../external-contact/add/emailpro-external-contact-add.html';
+import emailproExternalContactTemplate from '../external-contact/emailpro-external-contact.html';
 import emailproTemplate from './emailpro.html';
-import emailproTaskTemplate from './task/emailpro-task.html';
+import emailproTaskTemplate from '../task/emailpro-task.html';
 import emailproErrorMessagesPopoverTemplate from './emailpro-error-messages.popover.html';
-import emailproAccountOrderTemplate from './account/order/emailpro-account-order.html';
-import emailproAccountAliasRemoveTemplate from './account/alias/remove/emailpro-account-alias-remove.html';
-import emailproAccountAliasTemplate from './account/alias/emailpro-account-alias.html';
-import emailproAccountAliasAddTemplate from './account/alias/add/emailpro-account-alias-add.html';
-import emailproAccountRemoveTemplate from './account/remove/emailpro-account-remove.html';
-import emailproAccountUpdateTemplate from './account/update/emailpro-account-update.html';
-import emailproAccountDelegationTemplate from './account/delegation/emailpro-account-delegation.html';
-import emailproAccountAddTemplate from './account/add/emailpro-account-add.html';
-import emailproAccountTemplate from './account/emailpro-account.html';
-import emailproServiceConfigureTemplate from './service/configure/emailpro-service-configure.html';
-import emailproDomainMxAutoconfigTemplate from './domain/mx-autoconfig/emailpro-domain-mx-autoconfig.html';
-import emailproDomainTemplate from './domain/emailpro-domain.html';
-import emailproDomainRemoveTemplate from './domain/remove/emailpro-domain-remove.html';
-import emailproDomainUpdateTemplate from './domain/update/emailpro-domain-update.html';
-import emailproDomainSrvAutoconfigTemplate from './domain/srv-autoconfig/emailpro-domain-srv-autoconfig.html';
-import emailproDomainAddTemplate from './domain/add/emailpro-domain-add.html';
-import emailproMailingListUpdateTemplate from './mailing-list/update/emailpro-mailing-list-update.html';
-import emailproMailingListDeleteTemplate from './mailing-list/delete/emailpro-mailing-list-delete.html';
-import emailproMailingListSendListByEmailTemplate from './mailing-list/send-list-by-email/emailpro-mailing-list-send-list-by-email.html';
-import emailproMailingListModeratorDeleteTemplate from './mailing-list/moderator/delete/emailpro-mailing-list-moderator-delete.html';
-import emailproMailingListViewTemplate from './mailing-list/emailpro-mailing-list-view.html';
-import emailproMailingListTemplate from './mailing-list/emailpro-mailing-list.html';
-import emailproMailingListModeratorsDeleteTemplate from './mailing-list/moderators/delete/emailpro-mailing-list-moderators-delete.html';
-import emailproMailingListModeratorsViewTemplate from './mailing-list/moderators/emailpro-mailing-list-moderators-view.html';
-import emailproMailingListModeratorsCreateTemplate from './mailing-list/moderators/create/emailpro-mailing-list-moderators-create.html';
-import emailproMailingListSubscribersDeleteTemplate from './mailing-list/subscribers/delete/emailpro-mailing-list-subscribers-delete.html';
-import emailproMailingListSubscribersViewTemplate from './mailing-list/subscribers/emailpro-mailing-list-subscribers-view.html';
-import emailproMailingListSubscribersCreateTemplate from './mailing-list/subscribers/create/emailpro-mailing-list-subscribers-create.html';
-import emailproMailingListSubscriberDeleteTemplate from './mailing-list/subscriber/delete/emailpro-mailing-list-subscriber-delete.html';
-import emailproMailingListCreateTemplate from './mailing-list/create/emailpro-mailing-list-create.html';
+import emailproAccountOrderTemplate from '../account/order/emailpro-account-order.html';
+import emailproAccountAliasRemoveTemplate from '../account/alias/remove/emailpro-account-alias-remove.html';
+import emailproAccountAliasTemplate from '../account/alias/emailpro-account-alias.html';
+import emailproAccountAliasAddTemplate from '../account/alias/add/emailpro-account-alias-add.html';
+import emailproAccountRemoveTemplate from '../account/remove/emailpro-account-remove.html';
+import emailproAccountUpdateTemplate from '../account/update/emailpro-account-update.html';
+import emailproAccountDelegationTemplate from '../account/delegation/emailpro-account-delegation.html';
+import emailproAccountAddTemplate from '../account/add/emailpro-account-add.html';
+import emailproAccountTemplate from '../account/emailpro-account.html';
+import emailproServiceConfigureTemplate from '../service/configure/emailpro-service-configure.html';
+import emailproDomainMxAutoconfigTemplate from '../domain/mx-autoconfig/emailpro-domain-mx-autoconfig.html';
+import emailproDomainTemplate from '../domain/emailpro-domain.html';
+import emailproDomainRemoveTemplate from '../domain/remove/emailpro-domain-remove.html';
+import emailproDomainUpdateTemplate from '../domain/update/emailpro-domain-update.html';
+import emailproDomainSrvAutoconfigTemplate from '../domain/srv-autoconfig/emailpro-domain-srv-autoconfig.html';
+import emailproDomainAddTemplate from '../domain/add/emailpro-domain-add.html';
+import emailproMailingListUpdateTemplate from '../mailing-list/update/emailpro-mailing-list-update.html';
+import emailproMailingListDeleteTemplate from '../mailing-list/delete/emailpro-mailing-list-delete.html';
+import emailproMailingListSendListByEmailTemplate from '../mailing-list/send-list-by-email/emailpro-mailing-list-send-list-by-email.html';
+import emailproMailingListModeratorDeleteTemplate from '../mailing-list/moderator/delete/emailpro-mailing-list-moderator-delete.html';
+import emailproMailingListViewTemplate from '../mailing-list/emailpro-mailing-list-view.html';
+import emailproMailingListTemplate from '../mailing-list/emailpro-mailing-list.html';
+import emailproMailingListModeratorsDeleteTemplate from '../mailing-list/moderators/delete/emailpro-mailing-list-moderators-delete.html';
+import emailproMailingListModeratorsViewTemplate from '../mailing-list/moderators/emailpro-mailing-list-moderators-view.html';
+import emailproMailingListModeratorsCreateTemplate from '../mailing-list/moderators/create/emailpro-mailing-list-moderators-create.html';
+import emailproMailingListSubscribersDeleteTemplate from '../mailing-list/subscribers/delete/emailpro-mailing-list-subscribers-delete.html';
+import emailproMailingListSubscribersViewTemplate from '../mailing-list/subscribers/emailpro-mailing-list-subscribers-view.html';
+import emailproMailingListSubscribersCreateTemplate from '../mailing-list/subscribers/create/emailpro-mailing-list-subscribers-create.html';
+import emailproMailingListSubscriberDeleteTemplate from '../mailing-list/subscriber/delete/emailpro-mailing-list-subscriber-delete.html';
+import emailproMailingListCreateTemplate from '../mailing-list/create/emailpro-mailing-list-create.html';
 
 export default /* @ngInject */ ($templateCache) => {
   $templateCache.put(
