+++ packages/manager/modules/emailpro/src/dashboard/emailproServices.module.js
@@ -2,11 +2,11 @@ import angular from 'angular';
 import ovhManagerCore from '@ovh-ux/manager-core';
 
 import emailproService from './emailpro.service';
-import ApiEmailPro from './api/emailpro-api.service';
-import EmailProDomains from './domain/emailpro-domain.service';
-import EmailProExternalContacts from './external-contact/emailpro-external-contact.service';
-import EmailProPassword from './password/emailpro-password.service';
-import EmailproMxPlanMailingLists from './mailing-list/emailpro-mailing-list.service';
+import ApiEmailPro from '../api/emailpro-api.service';
+import EmailProDomains from '../domain/emailpro-domain.service';
+import EmailProExternalContacts from '../external-contact/emailpro-external-contact.service';
+import EmailProPassword from '../password/emailpro-password.service';
+import EmailproMxPlanMailingLists from '../mailing-list/emailpro-mailing-list.service';
 
 const moduleName = 'Module.emailpro.services';
