+++ packages/manager/modules/emailpro/src/dashboard/translations/Messages_es_ES.json
