+++ packages/manager/modules/emailpro/src/email-pro.routing.js
@@ -19,4 +19,23 @@ export default /* @ngInject */ ($stateProvider) => {
         }),
     },
   });
+
+  $stateProvider.state('app.mxplan', {
+    url: `/configuration/email_mxplan?${ListLayoutHelper.urlQueryParams}`,
+    component: 'managerListLayout',
+    params: ListLayoutHelper.stateParams,
+    resolve: {
+      ...ListLayoutHelper.stateResolves,
+      apiPath: () => '/email/mxplan',
+      dataModel: () => 'email.mxplan.Service',
+      defaultFilterColumn: () => 'domain',
+      header: /* @ngInject */ ($translate) =>
+        $translate.instant('emails_mx_plan_title'),
+      customizableColumns: () => true,
+      getServiceNameLink: /* @ngInject */ ($state) => ({ domain: productId }) =>
+        $state.href('app.email.mxplan', {
+          productId,
+        }),
+    },
+  });
 };
