--- packages/manager/modules/emailpro/src/email-pro/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "email_pro_title": "Emails Pro"
-}
