--- packages/manager/modules/emailpro/src/mx-plan/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "emails_mx_plan_title": "Emails"
-}
